Moves validation logic of ReturnsArgumentAt in the answer itself
